fix(control): centrilization of a fonction to avoid duplication
